requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
databases